+++ cumulus/client/consensus/aura/src/import_queue.rs
@@ -20,20 +20,183 @@
 use codec::Codec;
 use cumulus_client_consensus_common::ParachainBlockImportMarker;
 use prometheus_endpoint::Registry;
-use sc_client_api::{backend::AuxStore, BlockOf, UsageProvider};
-use sc_consensus::{import_queue::DefaultImportQueue, BlockImport};
-use sc_consensus_aura::{AuraVerifier, CompatibilityMode};
-use sc_consensus_slots::InherentDataProviderExt;
-use sc_telemetry::TelemetryHandle;
-use sp_api::{ApiExt, ProvideRuntimeApi};
+use sc_consensus::{
+	import_queue::{BasicQueue, Verifier as VerifierT},
+	BlockImport, BlockImportParams, ForkChoiceStrategy,
+};
+use sc_consensus_aura::standalone as aura_internal;
+use sc_telemetry::{telemetry, TelemetryHandle, CONSENSUS_DEBUG, CONSENSUS_TRACE};
+use sp_api::ProvideRuntimeApi;
 use sp_block_builder::BlockBuilder as BlockBuilderApi;
-use sp_blockchain::HeaderBackend;
-use sp_consensus::Error as ConsensusError;
-use sp_consensus_aura::AuraApi;
+use sp_consensus::error::Error as ConsensusError;
+use sp_consensus_aura::{AuraApi, Slot, SlotDuration};
 use sp_core::crypto::Pair;
-use sp_inherents::CreateInherentDataProviders;
-use sp_runtime::traits::Block as BlockT;
+use sp_inherents::{CreateInherentDataProviders, InherentDataProvider};
+use sp_runtime::traits::{Block as BlockT, Header as HeaderT};
 use std::{fmt::Debug, sync::Arc};
+
+/// A parachain block import verifier that checks for equivocation limits within each slot.
+pub struct Verifier<P, Client, Block, CIDP> {
+	client: Arc<Client>,
+	create_inherent_data_providers: CIDP,
+	telemetry: Option<TelemetryHandle>,
+	_phantom: std::marker::PhantomData<fn() -> (Block, P)>,
+}
+
+impl<P, Client, Block, CIDP> Verifier<P, Client, Block, CIDP>
+where
+	P: Pair,
+	P::Signature: Codec,
+	P::Public: Codec + Debug,
+	Block: BlockT,
+	Client: ProvideRuntimeApi<Block> + Send + Sync,
+	<Client as ProvideRuntimeApi<Block>>::Api: BlockBuilderApi<Block> + AuraApi<Block, P::Public>,
+
+	CIDP: CreateInherentDataProviders<Block, ()>,
+{
+	/// Creates a new Verifier instance for handling parachain block import verification in Aura
+	/// consensus.
+	pub fn new(
+		client: Arc<Client>,
+		inherent_data_provider: CIDP,
+		telemetry: Option<TelemetryHandle>,
+	) -> Self {
+		Self {
+			client,
+			create_inherent_data_providers: inherent_data_provider,
+			telemetry,
+			_phantom: std::marker::PhantomData,
+		}
+	}
+}
+
+#[async_trait::async_trait]
+impl<P, Client, Block, CIDP> VerifierT<Block> for Verifier<P, Client, Block, CIDP>
+where
+	P: Pair,
+	P::Signature: Codec,
+	P::Public: Codec + Debug,
+	Block: BlockT,
+	Client: ProvideRuntimeApi<Block> + Send + Sync,
+	<Client as ProvideRuntimeApi<Block>>::Api: BlockBuilderApi<Block> + AuraApi<Block, P::Public>,
+
+	CIDP: CreateInherentDataProviders<Block, ()>,
+{
+	async fn verify(
+		&self,
+		mut block_params: BlockImportParams<Block>,
+	) -> Result<BlockImportParams<Block>, String> {
+		// Skip checks that include execution, if being told so, or when importing only state.
+		//
+		// This is done for example when gap syncing and it is expected that the block after the gap
+		// was checked/chosen properly, e.g. by warp syncing to this block using a finality proof.
+		if block_params.state_action.skip_execution_checks() || block_params.with_state() {
+			block_params.fork_choice = Some(ForkChoiceStrategy::Custom(block_params.with_state()));
+			return Ok(block_params)
+		}
+
+		let post_hash = block_params.header.hash();
+		let parent_hash = *block_params.header.parent_hash();
+
+		// check seal and update pre-hash/post-hash
+		{
+			let authorities = aura_internal::fetch_authorities(self.client.as_ref(), parent_hash)
+				.map_err(|e| {
+				format!("Could not fetch authorities at {:?}: {}", parent_hash, e)
+			})?;
+
+			let slot_duration = self
+				.client
+				.runtime_api()
+				.slot_duration(parent_hash)
+				.map_err(|e| e.to_string())?;
+
+			let slot_now = slot_now(slot_duration);
+			let res = aura_internal::check_header_slot_and_seal::<Block, P>(
+				slot_now,
+				block_params.header,
+				&authorities,
+			);
+
+			match res {
+				Ok((pre_header, _, seal_digest)) => {
+					telemetry!(
+						self.telemetry;
+						CONSENSUS_TRACE;
+						"aura.checked_and_importing";
+						"pre_header" => ?pre_header,
+					);
+
+					block_params.header = pre_header;
+					block_params.post_digests.push(seal_digest);
+					block_params.fork_choice = Some(ForkChoiceStrategy::LongestChain);
+					block_params.post_hash = Some(post_hash);
+				},
+				Err(aura_internal::SealVerificationError::Deferred(hdr, slot)) => {
+					telemetry!(
+						self.telemetry;
+						CONSENSUS_DEBUG;
+						"aura.header_too_far_in_future";
+						"hash" => ?post_hash,
+						"a" => ?hdr,
+						"b" => ?slot,
+					);
+
+					return Err(format!(
+						"Rejecting block ({:?}) from future slot {:?}",
+						post_hash, slot
+					))
+				},
+				Err(e) =>
+					return Err(format!(
+						"Rejecting block ({:?}) with invalid seal ({:?})",
+						post_hash, e
+					)),
+			}
+		}
+
+		// check inherents.
+		if let Some(body) = block_params.body.clone() {
+			let block = Block::new(block_params.header.clone(), body);
+			let create_inherent_data_providers = self
+				.create_inherent_data_providers
+				.create_inherent_data_providers(parent_hash, ())
+				.await
+				.map_err(|e| format!("Could not create inherent data {:?}", e))?;
+
+			let inherent_data = create_inherent_data_providers
+				.create_inherent_data()
+				.await
+				.map_err(|e| format!("Could not create inherent data {:?}", e))?;
+
+			let inherent_res = self
+				.client
+				.runtime_api()
+				.check_inherents(parent_hash, block, inherent_data)
+				.map_err(|e| format!("Unable to check block inherents {:?}", e))?;
+
+			if !inherent_res.ok() {
+				for (i, e) in inherent_res.into_errors() {
+					match create_inherent_data_providers.try_handle_error(&i, &e).await {
+						Some(res) => res.map_err(|e| format!("Inherent Error {:?}", e))?,
+						None =>
+							return Err(format!(
+								"Unknown inherent error, source {:?}",
+								String::from_utf8_lossy(&i[..])
+							)),
+					}
+				}
+			}
+		}
+
+		Ok(block_params)
+	}
+}
+
+fn slot_now(slot_duration: SlotDuration) -> Slot {
+	let timestamp = sp_timestamp::InherentDataProvider::from_system_time().timestamp();
+	Slot::from_timestamp(timestamp, slot_duration)
+}
 
 /// Parameters for [`import_queue`].
 pub struct ImportQueueParams<'a, I, C, CIDP, S> {
@@ -50,7 +213,6 @@ pub struct ImportQueueParams<'a, I, C, CIDP, S> {
 	/// The telemetry handle.
 	pub telemetry: Option<TelemetryHandle>,
 }
-
 /// Start an import queue for the Aura consensus algorithm.
 pub fn import_queue<P, Block, I, C, S, CIDP>(
 	ImportQueueParams {
@@ -61,18 +223,11 @@ pub fn import_queue<P, Block, I, C, S, CIDP>(
 		registry,
 		telemetry,
 	}: ImportQueueParams<'_, I, C, CIDP, S>,
-) -> Result<DefaultImportQueue<Block>, sp_consensus::Error>
+) -> Result<BasicQueue<Block>, sp_consensus::Error>
 where
 	Block: BlockT,
-	C::Api: BlockBuilderApi<Block> + AuraApi<Block, P::Public> + ApiExt<Block>,
-	C: 'static
-		+ ProvideRuntimeApi<Block>
-		+ BlockOf
-		+ Send
-		+ Sync
-		+ AuxStore
-		+ UsageProvider<Block>
-		+ HeaderBackend<Block>,
+	C::Api: BlockBuilderApi<Block> + AuraApi<Block, P::Public>,
+	C: 'static + ProvideRuntimeApi<Block> + Send + Sync,
 	I: BlockImport<Block, Error = ConsensusError>
 		+ ParachainBlockImportMarker
 		+ Send
@@ -83,19 +238,15 @@ where
 	P::Signature: Codec,
 	S: sp_core::traits::SpawnEssentialNamed,
 	CIDP: CreateInherentDataProviders<Block, ()> + Sync + Send + 'static,
-	CIDP::InherentDataProviders: InherentDataProviderExt + Send + Sync,
 {
-	sc_consensus_aura::import_queue::<P, _, _, _, _, _>(sc_consensus_aura::ImportQueueParams {
-		block_import,
-		justification_import: None,
+	let verifier = Verifier::<P, _, _, _> {
 		client,
 		create_inherent_data_providers,
-		spawner,
-		registry,
-		check_for_equivocation: sc_consensus_aura::CheckForEquivocation::No,
 		telemetry,
-		compatibility_mode: CompatibilityMode::None,
-	})
+		_phantom: std::marker::PhantomData,
+	};
+
+	Ok(BasicQueue::new(verifier, Box::new(block_import), None, spawner, registry))
 }
 
 /// Parameters of [`build_verifier`].
@@ -108,18 +259,17 @@ pub struct BuildVerifierParams<C, CIDP> {
 	pub telemetry: Option<TelemetryHandle>,
 }
 
-/// Build the [`AuraVerifier`].
-pub fn build_verifier<P, C, CIDP, N>(
+/// Build the [`Verifier`].
+pub fn build_verifier<P, C, CIDP, Block>(
 	BuildVerifierParams { client, create_inherent_data_providers, telemetry }: BuildVerifierParams<
 		C,
 		CIDP,
 	>,
-) -> AuraVerifier<C, P, CIDP, N> {
-	sc_consensus_aura::build_verifier(sc_consensus_aura::BuildVerifierParams {
+) -> Verifier<P, C, Block, CIDP> {
+	Verifier {
 		client,
 		create_inherent_data_providers,
 		telemetry,
-		check_for_equivocation: sc_consensus_aura::CheckForEquivocation::No,
-		compatibility_mode: CompatibilityMode::None,
-	})
+		_phantom: std::marker::PhantomData,
+	}
 }
